IDKBROTHER

+++ src/kernel/memory.c
@@ -1,33 +1,33 @@
 #include "memory.h"
 #include <stdint.h>
 
-void* memcpy(void* dst, const void* src, uint16_t num)
+void* memcpy(void* dst, const void* src, size_t num)
 {
     uint8_t* u8Dst = (uint8_t *)dst;
     const uint8_t* u8Src = (const uint8_t *)src;
 
-    for (uint16_t i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
         u8Dst[i] = u8Src[i];
 
     return dst;
 }
 
-void * memset(void * ptr, int value, uint16_t num)
+void * memset(void * ptr, int value, size_t num)
 {
     uint8_t* u8Ptr = (uint8_t *)ptr;
 
-    for (uint16_t i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
         u8Ptr[i] = (uint8_t)value;
 
     return ptr;
 }
 
-int memcmp(const void* ptr1, const void* ptr2, uint16_t num)
+int memcmp(const void* ptr1, const void* ptr2, size_t num)
 {
     const uint8_t* u8Ptr1 = (const uint8_t *)ptr1;
     const uint8_t* u8Ptr2 = (const uint8_t *)ptr2;
 
-    for (uint16_t i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
     {
         if (u8Ptr1[i] != u8Ptr2[i])
             return u8Ptr1[i] - u8Ptr2[i];
@@ -36,18 +36,18 @@ int memcmp(const void* ptr1, const void* ptr2, uint16_t num)
     return 0;
 }
 
-void* memmove(void* dst, const void* src, uint16_t num)
+void* memmove(void* dst, const void* src, size_t num)
 {
     uint8_t* u8Dst = (uint8_t*)dst;
     const uint8_t* u8Src = (const uint8_t*)src;
 
     // If destination is before source, we can copy forwards
     if (u8Dst < u8Src) {
-        for (uint16_t i = 0; i < num; i++) {
+        for (size_t i = 0; i < num; i++) {
             u8Dst[i] = u8Src[i];
         }
     } else { // Otherwise, we must copy backwards to avoid overwriting data
-        for (uint16_t i = num; i > 0; i--) {
+        for (size_t i = num; i > 0; i--) {
             u8Dst[i-1] = u8Src[i-1];
         }
     }
